FIXED Max

+++ src/Utils/utils.java
@@ -63,7 +63,7 @@ public class utils {
 		}
 	}
 
-	public static double max(double a, double b){
+	public static int max(int a, int b){
 		if(a>b) return a;
 		else return b;
 	}
